backend/requirements.txt
asgiref
Django
psycopg2-binary
sqlparse
tzdata
whitenoise
gunicorn
djangorestframework
pandas